Moving boot order component to global
