+++ locales/en/plugin__kubevirt-plugin.json
@@ -1,4 +1,8 @@
 {
+  "No resource selected": "No resource selected",
+  "Virtual Machine will attempt to boot from disks by order of appearance in YAML file": "Virtual Machine will attempt to boot from disks by order of appearance in YAML file",
+  "Hide default boot disks": "Hide default boot disks",
+  "Show default boot disks": "Show default boot disks",
   "Edit CPU | Memory": "Edit CPU | Memory",
   "Save": "Save",
   "Restore default settings": "Restore default settings",
@@ -144,10 +148,6 @@
   "Events": "Events",
   "Console": "Console",
   "Snapshots": "Snapshots",
-  "No resource selected": "No resource selected",
-  "Virtual Machine will attempt to boot from disks by order of appearance in YAML file": "Virtual Machine will attempt to boot from disks by order of appearance in YAML file",
-  "Hide default boot disks": "Hide default boot disks",
-  "Show default boot disks": "Show default boot disks",
   "None": "None",
   "Name must be unique within a namespace. Is required when creating resources, although some resources may allow a client to request the generation of an appropriate name automatically. Name is primarily intended for creation idempotence and configuration definition. Cannot be updated. ": "Name must be unique within a namespace. Is required when creating resources, although some resources may allow a client to request the generation of an appropriate name automatically. Name is primarily intended for creation idempotence and configuration definition. Cannot be updated. ",
   "Namespace defines the space within which each name must be unique. An empty namespace is equivalent to the \"default\" namespace, but \"default\" is the canonical representation. Not all objects are required to be scoped to a namespace - the value of this field for those objects will be empty. Must be a DNS_LABEL. Cannot be updated. ": "Namespace defines the space within which each name must be unique. An empty namespace is equivalent to the \"default\" namespace, but \"default\" is the canonical representation. Not all objects are required to be scoped to a namespace - the value of this field for those objects will be empty. Must be a DNS_LABEL. Cannot be updated. ",
